Further reduce candidate map
Check whether candidates for a number in a line or row are confined to a subsquare. If so, the rest of that subsquare cannot be a candidate for the number. Likewise, if candidates for a number in a subsquare are confined to a line (or row), the remaining vertices of that line (or row) are not candidates for the number.